feat: implement shared MessageView widget for consistent message display across screens
