--- mobile/lib/features/streams/presentation/widgets/message_view.dart
@@ -1,46 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Message pleine hauteur (icône + texte + action optionnelle), scrollable pour
-/// rester compatible avec un `RefreshIndicator`. Partagé par les listes de flux.
-class MessageView extends StatelessWidget {
-  const MessageView({
-    super.key,
-    required this.icon,
-    required this.message,
-    this.actionLabel,
-    this.onAction,
-  });
-
-  final IconData icon;
-  final String message;
-  final String? actionLabel;
-  final VoidCallback? onAction;
-
-  @override
-  Widget build(BuildContext context) {
-    final text = Theme.of(context).textTheme;
-    final colors = Theme.of(context).colorScheme;
-
-    return ListView(
-      children: [
-        SizedBox(height: MediaQuery.of(context).size.height * 0.25),
-        Icon(icon, size: 64, color: colors.onSurfaceVariant),
-        const SizedBox(height: 16),
-        Text(
-          message,
-          textAlign: TextAlign.center,
-          style: text.titleMedium?.copyWith(color: colors.onSurfaceVariant),
-        ),
-        if (actionLabel != null && onAction != null) ...[
-          const SizedBox(height: 16),
-          Center(
-            child: FilledButton(
-              onPressed: onAction,
-              child: Text(actionLabel!),
-            ),
-          ),
-        ],
-      ],
-    );
-  }
-}
